create WalletConnectV2Account.kt to make xmtp SDK with WC v2
